Fix crash loading conformance metadata: vendor command IDs can exceed ulong
AuthenticatorGetInfo.VendorPrototypeConfigCommands was int[]; the FIDO
conformance tool's test metadata sends values up to 2^64 to exercise this
field (one of its 25 exported statements has [18446744073709552000,
184467440737095520]), which overflows int, long and even ulong. That one
statement's JsonException took down the entire directory load, so every
attestation test failed with AaGuidNotFound regardless of which authenticator
it used -- not just the one exercising this field.

Widened to double[] (subcommand identifiers are otherwise always small, and
round-trip through double exactly). Also hardened
FileSystemMetadataRepository.GetBLOBAsync so one statement file with an
unexpected shape is skipped rather than failing every other statement in the
directory, the same way the rest of this branch already treats malformed
input from an outside source as a rejection of that input rather than a
crash of the whole operation.

+++ Src/Fido2.Models/Metadata/AuthenticatorGetInfo.cs
@@ -141,10 +141,13 @@ public sealed class AuthenticatorGetInfo
     public int? RemainingDiscoverableCredentials { get; set; }
 
     /// <summary>
-    /// List of authenticatorConfig vendorPrototype subcommand identifiers.
+    /// List of authenticatorConfig vendorPrototype subcommand identifiers. Real subcommand identifiers are small,
+    /// but this is <see langword="double"/> rather than <see langword="int"/> because the FIDO conformance tool's
+    /// test metadata deliberately sends values as large as 2^64 to exercise this field, which int/long/ulong all
+    /// overflow on.
     /// </summary>
     [JsonPropertyName("vendorPrototypeConfigCommands")]
-    public int[]? VendorPrototypeConfigCommands { get; set; }
+    public double[]? VendorPrototypeConfigCommands { get; set; }
 
     /// <summary>
     /// List of supported attestation statement formats, in decreasing order of authenticator preference.
